+++ editoast/editoast_derive/src/modelv2/codegen/create_impl.rs
@@ -1,0 +1,39 @@
+use quote::quote;
+use quote::ToTokens;
+
+pub(crate) struct CreateImpl {
+    pub(super) model: syn::Ident,
+    pub(super) table_mod: syn::Path,
+    pub(super) row: syn::Ident,
+    pub(super) changeset: syn::Ident,
+}
+
+impl ToTokens for CreateImpl {
+    fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
+        let Self {
+            model,
+            table_mod,
+            row,
+            changeset,
+        } = self;
+
+        tokens.extend(quote! {
+            #[automatically_derived]
+            #[async_trait::async_trait]
+            impl crate::modelsv2::Create<#model> for #changeset {
+                async fn create(
+                    self,
+                    conn: &mut diesel_async::AsyncPgConnection,
+                ) -> crate::error::Result<#model> {
+                    use diesel_async::RunQueryDsl;
+                    diesel::insert_into(#table_mod::table)
+                        .values(&self)
+                        .get_result::<#row>(conn)
+                        .await
+                        .map(Into::into)
+                        .map_err(Into::into)
+                }
+            }
+        });
+    }
+}
